[ch07]-[block-out2]のhiwordマクロとlowordマクロの型をi32->i16に修正

+++ ch07/block-out2/src/macros.rs
@@ -1,11 +1,11 @@
 #[allow(unused_macros)]
 macro_rules! hiword {
-	($param:expr) => { (($param >> 16) & 0xFFFF) as i32};
+	($param:expr) => { (($param >> 16) & 0xFFFF) as i16};
 }
 
 #[allow(unused_macros)]
 macro_rules! loword {
-	($param:expr) => {($param & 0xFFFF) as i32};
+	($param:expr) => {($param & 0xFFFF) as i16};
 }
 
 #[allow(unused_macros)]
